pyproject.toml
[project]
name = "anthemai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.203.0",
    "crewai-tools>=0.76.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
